Cleanup generate_rollouts according to functools partial and move utils into alg_utils
