+++ alg_utils.py
@@ -10,6 +10,7 @@ from freezable_policy import FreezableModularPolicy
 from ray.tune.logger import UnifiedLogger
 
 from functools import partial
+from copy import deepcopy
 
 from tictactoe_wrapper import get_other
 
@@ -173,3 +174,43 @@ def get_policies(policies, obs_space, act_space, dim):
 
 def get_single_trainer(trainer, pols, ind, env, logdir):
     return trainer(policy_mapping_fn(ind), pols, env, logdir)
+
+
+def get_pretrained_weight_subset(trainer, index, pretrained_weight_keywords):
+    def any_keyword_in_key(keywords, key):
+        for kw in keywords:
+            if kw in key:
+                return True
+        return False
+
+    assert pretrained_weight_keywords is not None
+    policy_key = policy_mapping_fn(index)
+
+    # Pull out initial weights, and remove those we want to re-initialize rather than use pretrained values for
+    initial_weights = deepcopy(trainer.get_weights(policy_key))[policy_key]
+    initial_weights = {k: v for k, v in initial_weights.items()
+                       if any_keyword_in_key(pretrained_weight_keywords, k)}
+
+    return {policy_key: initial_weights}
+
+
+def possibly_pretrained_trainer(trainer_func, ind, policies, env, logdir, checkpoint_path=None,
+                                load_weights=True, pretrained_weight_keywords=None):
+    # If load_weights=True pretrained_weight_keywords is None, all weights are kept
+    # If load_weights=False, a fully reinitialized trainer is returned
+    trainer = get_single_trainer(trainer_func, policies, ind, env, logdir)
+    initial_random_weights = trainer.get_weights(policy_mapping_fn(ind))
+
+    if not load_weights:
+        # if we're not loading any weights, return a randomly initialized trainer
+        return trainer
+
+    trainer.restore(checkpoint_path)
+    if pretrained_weight_keywords is None:
+        # if we're not subsetting out specific weights to keep, then don't calculate a weight mask
+        return trainer
+
+    weight_mask = get_pretrained_weight_subset(trainer, ind, pretrained_weight_keywords)
+    trainer.set_weights(initial_random_weights)
+    trainer.set_weights(weight_mask)
+    return trainer
