+++ generate_rollouts.py
@@ -1,39 +1,41 @@
 from sacred import Experiment
-from sacred.observers import FileStorageObserver
-
-from tictactoe_wrapper import MultiAgentTicTacToe
 from marl_training import policy_mapping_fn, sync_trainers, simulate_rollouts_tictactoe
 from env_utils import get_env
 import ray
-from alg_utils import get_algs
+from alg_utils import get_pg_pol, get_pg_train, get_policies, possibly_pretrained_trainer, get_hierarchical_pg_pol
 from measure_exploitability import _get_checkpoint_from_run
-
+from functools import partial
 
 
 ex = Experiment("simulate_rollouts")
 
 @ex.config
 def my_config():
-    agent1_alg = "pg"
-    agent2_alg = "pg_hierarchical_1by256"
+    gamma = 0.95
+    prev_run = None
+    dir_loc = None
+    agent_algorithms = [get_pg_pol,  partial(get_hierarchical_pg_pol,
+                                             core_policy_hiddens=(256,),
+                                             redundant_policy_hiddens=(256,))]
+    agent_trainers = [partial(get_pg_train, gamma=gamma), partial(get_pg_train, gamma=gamma)]
     env = "tictactoe-v0"
     dim = 10
     training_iters = 200
     simulate_trained_pols = True
     silent = True
-    gamma = 0.95
     checkpoint_freq = 50
 
 @ex.automain
-def my_main(_run, agent1_alg, agent2_alg, dim, env, simulate_trained_pols, target, dir_loc):
+def my_main(_run, agent_algorithms, agent_trainers, dim, env, simulate_trained_pols, prev_run, dir_loc):
     ray.init()
     obs_space, act_space, env_name = get_env(env, dim)
-
-    trainers, pols = get_algs([agent1_alg, agent2_alg], obs_space, act_space, dim, env_name, "~/raydir/")
-
-    trainers[0].restore(_get_checkpoint_from_run(target, 0, dir_loc))
-    trainers[1].restore(_get_checkpoint_from_run(target, 1, dir_loc))
-
+    policies = get_policies(agent_algorithms, obs_space, act_space, dim)
+    checkpoint_paths = [_get_checkpoint_from_run(ind, prev_run, dir_loc) for ind in range(2)]
+    logdir = "~/raydir/" # change this
+    trainers = [
+        possibly_pretrained_trainer(agent_trainers[ind], ind, policies, env, logdir, checkpoint_paths[ind])
+        for ind in range(2)
+    ]
     sync_trainers(trainers)
 
     if simulate_trained_pols:
