+++ measure_exploitability.py
@@ -8,7 +8,7 @@ from ray.tune.logger import pretty_print
 
 from env_utils import get_env
 
-from alg_utils import get_single_trainer, policy_mapping_fn, get_pg_train, get_pg_pol, get_hierarchical_pg_pol, get_policies
+from alg_utils import get_single_trainer, policy_mapping_fn, get_pg_train, get_pg_pol, get_hierarchical_pg_pol, get_policies, possibly_pretrained_trainer
 
 from marl_training import sync_trainers
 from marl_training import SAVE_DIR as TRAINING_SAVE_DIR
@@ -297,48 +297,10 @@ def ordered_trainers(victim_index, adv_index, victim_trainer, adversary_trainer)
     return order_by_agent(victim_index, adv_index, victim_trainer, adversary_trainer)
 
 
-def get_pretrained_weight_subset(trainer, index, pretrained_weight_keywords):
-    def any_keyword_in_key(keywords, key):
-        for kw in keywords:
-            if kw in key:
-                return True
-        return False
 
-        # Figure out which string key corresponds to the policy in the adversarial position
-    assert pretrained_weight_keywords is not None
-    policy_key = policy_mapping_fn(index)
-
-    # Pull out initial weights, and remove those we want to re-initialize rather than use pretrained values for
-    initial_weights = deepcopy(trainer.get_weights(policy_key))[policy_key]
-    initial_weights = {k: v for k, v in initial_weights.items()
-                       if any_keyword_in_key(pretrained_weight_keywords, k)}
-
-    return {policy_key: initial_weights}
-
-
-def possibly_pretrained_trainer(trainer_func, ind, policies, env, logdir, checkpoint_path=None,
-                                load_weights=True, pretrained_weight_keywords=None):
-    # If load_weights=True pretrained_weight_keywords is None, all weights are kept
-    # If load_weights=False, a fully reinitialized trainer is returned
-    trainer = get_single_trainer(trainer_func, policies, ind, env, logdir)
-    initial_random_weights = trainer.get_weights(policy_mapping_fn(ind))
-
-    if not load_weights:
-        # if we're not loading any weights, return a randomly initialized trainer
-        return trainer
-
-    trainer.restore(checkpoint_path)
-    if pretrained_weight_keywords is None:
-        # if we're not subsetting out specific weights to keep, then don't calculate a weight mask
-        return trainer
-
-    weight_mask = get_pretrained_weight_subset(trainer, ind, pretrained_weight_keywords)
-    trainer.set_weights(initial_random_weights)
-    trainer.set_weights(weight_mask)
-    return trainer
 
 @exploitability_ex.automain
-def my_main(_run, dim, env,prev_run, victim_index, adv_index, outer_loop_iters, adversary_training_iters, victim_finetuning_iters,
+def my_main(_run, dim, env, prev_run, victim_index, adv_index, outer_loop_iters, adversary_training_iters, victim_finetuning_iters,
             victim_trainer_func, adversary_trainer_func, dir_loc, reload_checkpoint_index):
     ray.init()
     logdir = EXPLOIT_SAVE_DIR + "/" + str(_run._id) + "/rllib/"
